pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aden"
version = "0.2.0"
description = "Multi-provider LLM SDK metering - OpenAI, Anthropic, Gemini - usage tracking, budget enforcement, and cost control"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Aden", email = "aden@example.com" }
]
keywords = ["openai", "anthropic", "gemini", "llm", "metering", "cost", "usage", "billing", "erp", "livekit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# No required dependencies - all providers are optional
dependencies = []

[project.optional-dependencies]
# Individual provider support
openai = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
]
anthropic = [
    "anthropic>=0.20.0",
]
gemini = [
    "google-generativeai>=0.4.0",
]
genai = [
    "google-genai>=1.0.0",
]
# Google ADK (uses google-genai internally)
adk = [
    "google-adk>=0.1.0",
    "google-genai>=1.0.0",
]
# All providers
all = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "anthropic>=0.20.0",
    "google-generativeai>=0.4.0",
    "google-genai>=1.0.0",
]
# PydanticAI framework
pydantic-ai = [
    "pydantic-ai>=0.0.1",
    "openai>=1.0.0",
]
# LiveKit integration (requires OpenAI)
livekit = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "livekit-agents>=0.8.0",
    "livekit-plugins-openai>=0.8.0",
]
# Database backends
timescale = [
    "asyncpg>=0.29.0",
]
timescale-sync = [
    "psycopg2-binary>=2.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/aden/aden-py"
Documentation = "https://github.com/aden/aden-py#readme"
Repository = "https://github.com/aden/aden-py"

[tool.hatch.build.targets.wheel]
packages = ["src/aden"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
